fix(collections): fixes stale hubs data creating collections
ensures hub data is deleted when deleting a collection when becoming inactive due to time
restrictions. also removes incorrect logic which caused a pre-existing collection to be created if
it was only found it hub data, which is incorrect as collections remain in plex hubs even if the
collection is deleted in the library.